+++ packages/ui/src/index.ts
@@ -15,12 +15,6 @@ export type { OrderSummaryProps } from './components/molecules/OrderSummary'
 export { default as Bullets } from './components/molecules/Bullets'
 export type { BulletsProps } from './components/molecules/Bullets'
 
-export { default as SearchInput } from './components/molecules/SearchInput'
-export type {
-  SearchInputProps,
-  SearchInputRef,
-} from './components/molecules/SearchInput'
-
 export { default as Carousel } from './components/molecules/Carousel'
 export type { CarouselProps } from './components/molecules/Carousel'
 
@@ -41,17 +35,6 @@ export { default as Form } from './components/molecules/Form'
 export type { FormProps } from './components/molecules/Form'
 
 // Organisms
-export {
-  default as OutOfStock,
-  OutOfStockTitle,
-  OutOfStockMessage,
-} from './components/organisms/OutOfStock'
-export type {
-  OutOfStockProps,
-  OutOfStockMessageProps,
-  OutOfStockTitleProps,
-} from './components/organisms/OutOfStock'
-
 export { Tiles, Tile } from './components/organisms/Tiles'
 export type { TilesProps, TileProps } from './components/organisms/Tiles'